FINISH USER CRUD CONTROLLER

+++ app/Http/Controllers/UserController.php
@@ -2,7 +2,8 @@
 
 namespace App\Http\Controllers;
 
-use App\Http\Requests\UserStoreRequest;
+use App\Http\Requests\User\UserStoreRequest;
+use App\Http\Requests\User\UserUpdateRequest;
 use App\Http\Resources\UserCollection;
 use App\Http\Resources\UserResource;
 use App\Models\User;
@@ -19,6 +20,7 @@ class UserController extends Controller
      */
     public function index()
     {
+        $this->authorize('view_user');
         $users = User::all();
 
         return new UserCollection($users);
@@ -32,6 +34,7 @@ class UserController extends Controller
      */
     public function store(UserStoreRequest $request)
     {
+        $this->authorize('create_user');
         $user = new User;
         $user->name = $request->name;
         $user->email = $request->email;
@@ -42,15 +45,9 @@ class UserController extends Controller
 
         if ($user->save()) {
             $user->assignRole($request->role);
-            return response()->json([
-                'status' => true,
-                "message" => "User Created Successfully",
-            ]);
+            return store_message('User');
         } else {
-            return response()->json([
-                'status' => false,
-                "message" => "Please Try Again !",
-            ], 403);
+            return try_again_message();
         }
     }
 
@@ -62,6 +59,7 @@ class UserController extends Controller
      */
     public function show(User $user)
     {
+        $this->authorize('update_user');
         return new UserResource($user);
     }
 
@@ -70,11 +68,29 @@ class UserController extends Controller
      *
      * @param  \Illuminate\Http\Request  $request
      * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\Http\JsonResponse
      */
-    public function update(Request $request, User $user)
+    public function update(UserUpdateRequest $request, User $user)
     {
-        //
+        $this->authorize('update_user');
+
+        $user->name = $request->name;
+        $user->email = $request->email;
+        $user->role_id = $request->role;
+        ($request->password != "") ?  $user->password = Hash::make($request->password) : "";
+        $user->updated_by = \auth()->id();
+        $user->updated_at = Carbon::now();
+
+
+
+        if ($user->save()) {
+            $user->syncRoles($request->role);
+            return update_message('User');
+        } else {
+            return try_again_message();
+        }
+
+
     }
 
     /**
@@ -85,6 +101,7 @@ class UserController extends Controller
      */
     public function destroy(User $user)
     {
+        $this->authorize('delete_user');
         return $user->delete()
             ? response()->json([
                 'status' => true,
